# src/main/java/com/user/management/dao/AuthenticationResponse.java
package com.user.management.dao;

public class AuthenticationResponse {
	
	private final String jwt;

	public AuthenticationResponse(String jwt) {
		this.jwt = jwt;
	}

	public String getJwt() {
		return jwt;
	}
}